# docs/api/openapi.v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "V1Envelope": {
        "properties": {
          "data": {
            "description": "Resource payload; a list for collections.",
            "nullable": true
          },
          "errors": {
            "items": {
              "properties": {
                "code": {
                  "type": "string"
                },
                "detail": {
                  "type": "string"
                }
              },
              "type": "object"
            },
            "type": "array"
          },
          "meta": {
            "properties": {
              "api_version": {
                "type": "string"
              },
              "count": {
                "type": "integer"
              },
              "default_page": {
                "description": "True when the caller sent no limit, offset, or cursor and the route served its first default-size page; follow next_cursor for the rest.",
                "type": "boolean"
              },
              "limit": {
                "type": "integer"
              },
              "next_cursor": {
                "description": "Opaque cursor for the next page; null on the last page.",
                "nullable": true,
                "type": "string"
              },
              "offset": {
                "type": "integer"
              },
              "parts": {
                "description": "Per-list count and returned rows when an object payload is paged.",
                "type": "object"
              },
              "resource": {
                "type": "string"
              },
              "returned": {
                "type": "integer"
              }
            },
            "type": "object"
          }
        },
        "required": ["data", "meta", "errors"],
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Open-source, headless-first trust operations for customer-owned evidence lakes. The same engine serves GRC Lake Console and `/api/v1` for agents, CI, and runbooks. Discover the contract at `GET /api/v1`; browse it at `/docs`.",
    "title": "GRC Lake API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/operations": {
      "get": {
        "operationId": "operation_list_api_v1_operations_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Envelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Operation List",
        "tags": ["platform"]
      }
    },
...
    "/api/v1/platform/scim/tokens": {
      "get": {
        "operationId": "list_tokens_api_v1_platform_scim_tokens_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Envelope"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Tokens",
        "tags": ["commercial"]
      },
      "post": {
        "operationId": "create_token_api_v1_platform_scim_tokens_post",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Envelope"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Token",
        "tags": ["commercial"]
      }
    },
    "/api/v1/platform/scim/tokens/{token_id}": {
      "delete": {
        "operationId": "revoke_token_api_v1_platform_scim_tokens__token_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "title": "Token Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Envelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Token",
        "tags": ["commercial"]
      }
    },
...
  },
...
}